docker/services/voice-services/whisper-stt/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
pydantic==2.5.0
python-json-logger==2.0.7
